+++ src/simulation/Sink.java
@@ -6,8 +6,7 @@ import java.util.ArrayList;
  *	@author Joel Karel
  *	@version %I%, %G%
  */
-public class Sink implements ProductAcceptor
-{
+public class Sink implements ProductAcceptor {
 	/** All products are kept */
 	private ArrayList<Product> products;
 	/** All properties of products are kept */
@@ -23,8 +22,7 @@ public class Sink implements ProductAcceptor
 	/**
 	*	Constructor, creates objects
 	*/
-	public Sink(String n)
-	{
+	public Sink(String n) {
 		name = n;
 		products = new ArrayList<>();
 		numbers = new ArrayList<>();
@@ -34,17 +32,15 @@ public class Sink implements ProductAcceptor
 		number = 0;
 	}
 	
-        @Override
-	public boolean giveProduct(Product p)
-	{
+	@Override
+	public boolean giveProduct(Product p) {
 		number++;
 		products.add(p);
 		// store stamps
 		ArrayList<Double> t = p.getTimes();
 		ArrayList<String> e = p.getEvents();
 		ArrayList<String> s = p.getStations();
-		for(int i=0;i<t.size();i++)
-		{
+		for (int i = 0; i < t.size(); i++) {
 			numbers.add(number);
 			times.add(t.get(i));
 			events.add(e.get(i));
@@ -53,37 +49,31 @@ public class Sink implements ProductAcceptor
 		return true;
 	}
 	
-	public int[] getNumbers()
-	{
+	public int[] getNumbers() {
 		numbers.trimToSize();
 		int[] tmp = new int[numbers.size()];
-		for (int i=0; i < numbers.size(); i++)
-		{
-			tmp[i] = (numbers.get(i)).intValue();
+		for (int i = 0; i < numbers.size(); i++) {
+			tmp[i] = numbers.get(i);
 		}
 		return tmp;
 	}
 
-	public double[] getTimes()
-	{
+	public double[] getTimes() {
 		times.trimToSize();
 		double[] tmp = new double[times.size()];
-		for (int i=0; i < times.size(); i++)
-		{
-			tmp[i] = (times.get(i)).doubleValue();
+		for (int i = 0; i < times.size(); i++) {
+			tmp[i] = times.get(i);
 		}
 		return tmp;
 	}
 
-	public String[] getEvents()
-	{
+	public String[] getEvents() {
 		String[] tmp = new String[events.size()];
 		tmp = events.toArray(tmp);
 		return tmp;
 	}
 
-	public String[] getStations()
-	{
+	public String[] getStations() {
 		String[] tmp = new String[stations.size()];
 		tmp = stations.toArray(tmp);
 		return tmp;
